+++ Makefile
@@ -6,7 +6,7 @@
 #    By: nnakarac <nnakarac@42.fr>                  +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/10/09 01:24:09 by nnakarac          #+#    #+#              #
-#    Updated: 2023/02/05 19:32:51 by nnakarac         ###   ########.fr        #
+#    Updated: 2023/02/05 19:39:05 by nnakarac         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -91,6 +91,8 @@ SRCS	=	test.c \
 			ms_builtin_cmd.c \
 			ms_builtin_cmd_parent.c \
 			ms_builtin_cmd_child.c \
+			ms_builtin_cmd_utils.c \
+			ms_builtin_cmd_utils2.c \
 			ms_signals.c \
 			ms_debug.c \
 			ms_err.c \
