+++ srcs/ms_builtin_cmd_child.c
@@ -12,50 +12,6 @@
 
 #include "ms_input.h"
 
-int	cmd_bltn_pwd(t_cmd_lst *cmd)
-{
-	char	*pwd;
-
-	(void) cmd;
-	pwd = get_pwd();
-	write(1, pwd, ft_strlen(pwd));
-	write(1, "\n", 1);
-	return (0);
-}
-
-int	cmd_bltn_echo(t_cmd_lst *cmd)
-{
-	int	i;
-	int	is_newline;
-
-	i = 1;
-	is_newline = 1;
-	if (cmd->bargv[1])
-	{
-		while (!ft_strncmp(cmd->bargv[i], "-n\0", 3))
-		{
-			i++;
-			is_newline = 0;
-		}
-	}
-	while (cmd->bargv[i])
-	{
-		ft_putstr_fd(cmd->bargv[i++], 1);
-		if (cmd->bargv[i])
-			ft_putstr_fd(" ", 1);
-	}
-	if (is_newline)
-		ft_putstr_fd("\n", 1);
-	return (0);
-}
-
-int	cmd_bltn_env(t_cmd_lst *cmd)
-{
-	(void) cmd;
-	print_env();
-	return (0);
-}
-
 int	cmd_bltn_child_exec(t_cmd_lst *cmd)
 {
 	if (!ft_strncmp(cmd->bargv[0], "pwd\0", 4))
@@ -65,4 +21,14 @@ int	cmd_bltn_child_exec(t_cmd_lst *cmd)
 	if (!ft_strncmp(cmd->bargv[0], "env\0", 4))
 		return (cmd_bltn_env(cmd));
 	return (0);
+}
+
+char	**get_bltn_child(t_cmd_lst *cmd, char **bltn, int i, char **metas)
+{
+	(void) bltn;
+	cmd->btype = 0;
+	get_cmd_bargv(cmd, i, metas);
+	cmd->blen = arr2dsize(&metas[i]);
+	cmd_bltn_markout(cmd, i, metas);
+	return (cmd->bargv);
 }
