+++ srcs/ms_builtin_cmd_parent.c
@@ -6,44 +6,11 @@
 /*   By: nnakarac <nnakarac@42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/22 01:02:12 by nnakarac          #+#    #+#             */
-/*   Updated: 2023/02/05 18:16:54 by nnakarac         ###   ########.fr       */
+/*   Updated: 2023/02/05 19:44:34 by nnakarac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ms_input.h"
-
-int	cmd_bltn_cd(t_cmd_lst *cmd)
-{
-	if (cmd->bargv[1])
-		return (ms_cd(cmd->bargv[1]));
-	else
-		return (ms_cd(""));
-}
-
-int	cmd_bltn_exit(t_cmd_lst *cmd)
-{
-	int	exit_status;
-
-	exit_status = 0;
-	if (cmd->bargv[1])
-	{
-		if (!is_string_numeric(cmd->bargv[1]))
-		{
-			exit_status = ft_atoi(cmd->bargv[1]) % 256;
-			exit(exit_status);
-		}
-		else
-			err_exit(cmd->bargv[1], 1);
-	}
-	return (0);
-}
-
-int	cmd_bltn_unset(t_cmd_lst *cmd)
-{
-	if (cmd->bargv[1])
-		unset_env(cmd->bargv[1]);
-	return (0);
-}
 
 int	cmd_bltn_parent_exec(t_cmd_lst *cmd)
 {
@@ -54,8 +21,16 @@ int	cmd_bltn_parent_exec(t_cmd_lst *cmd)
 	if (!ft_strncmp(cmd->bargv[0], "unset\0", 6))
 		return (cmd_bltn_unset(cmd));
 	if (!ft_strncmp(cmd->bargv[0], "export\0", 7))
-	{
-
-	}
+		return (cmd_bltn_unset(cmd));
 	return (-1);
+}
+
+char	**get_bltn_parent(t_cmd_lst *cmd, char **bltn, int i, char **metas)
+{
+	(void) bltn;
+	cmd->btype = 1;
+	get_cmd_bargv(cmd, i, metas);
+	cmd->blen = arr2dsize(&metas[i]);
+	cmd_bltn_markout(cmd, i, metas);
+	return (cmd->bargv);
 }
