+++ srcs/ms_exec.c
@@ -6,7 +6,7 @@
 /*   By: nnakarac <nnakarac@42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/13 21:01:52 by nnakarac          #+#    #+#             */
-/*   Updated: 2023/02/05 19:24:40 by nnakarac         ###   ########.fr       */
+/*   Updated: 2023/02/05 19:38:39 by nnakarac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,6 @@ void	parent(t_cmd_lst *cmd, int *fd)
 
 void	exec(t_cmd_lst *cmd, int *fd)
 {
-	struct sigaction	s_int;
-
 	(void) fd;
 	if (cmd->pid)
 		cmd->pid = fork();
